requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.4
cryptography==41.0.1
databases==0.6.2
exceptiongroup==1.1.2
fastapi==0.100.0
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
orjson==3.9.2
ormar==0.12.2
passlib==1.7.4
psycopg2==2.9.6
pycparser==2.21
pydantic==1.10.8
pydantic_core==2.1.2
PyJWT==2.7.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.27.0
typer==0.9.0
typing_extensions==4.7.1
uvicorn==0.22.0
